requirements.txt
# DeepSight AI - Requirements
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
streamlit>=1.28.0
plotly>=5.15.0
tqdm>=4.65.0
albumentations>=1.3.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
mediapipe>=0.10.0
insightface>=0.7.0
